rename auth directory to middleware
